Fix most of hash-types re-exports
